Better wa to call Hook()
Allows using for example
`Hook( FUNC_PLAYER_KEYINPUT, &Player_keyInput_H, &Player_keyInput );`
instead of
`Hook( reinterpret_cast<void*>(FUNC_PLAYER_KEYINPUT), reinterpret_cast<void*>(&Player_keyInput_H), reinterpret_cast<void**>(&Player_keyInput) );` 

Further, this is not only easier to read and write, it does proper type-checking. It checks that the second and third arguments are a functionpointer and a pointer to a functionpointer, and that the signatures of the functions match (i.e. you are hooking a function with a matching function)